+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 Notable changes to this project.
 
-## dev
+## [2.9.3] - 2021-11-12
 - cli: support uploading predictions generated with the new dataset (needs `--new_data`)
 - signals: make `round_model_performances` available